move around operator code
